Fix bug in thermal_offset()
Thermal_offset() returned an offset rather than doing the full math in the datasheet, leading to an issue with order of operations. The return value was getting subtracted from the pixel data.
Formula in datasheet is Vij - [gradient] -
But If the values are 100, 50, 20 then result should be 30 (100 - 50 - 20)
Iff we return 50 - 20 (30) and that's later subtracted from 100 we get 70, which is incorrect.
 Instead we treturn (50 * -1 - 20) * -1 = 70. 100 - 70 = 30